getaddrinfo_async: Check return value of getservbyname_r earlier
If getservbyname_r fails (or HAVE_GETSERVBYNAME_R_4_ARGS is not defined)
it doesn't make any sense to access member of the servent struct as it
will not be initialized then. Calling endservent_r shouldn't be
necessary either then.

GCC 8 also complains about this code part when compiling with
the -Wmaybe-uninitialized flag.