Fix libtool link warning
Libtool produces a warning if we try to link a static lib into a shared lib (or any other .lx lib for that matter). The problem is not the linking which is fine since openbsd-compat uses -fPIC. The problem is that libtool doesn't know openbsd-compat is a convenience lib unless we use the proper suffix hence the warning.